moved report/helpful to helpers folder
